Move file views into files app
