+++ apps/documents/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 
 from apps.documents.models import Collection
-from apps.files.models import File
 from apps.service_providers.models import LlmProviderTypes
 
 
@@ -37,16 +36,3 @@ class CollectionForm(forms.ModelForm):
         if self.cleaned_data["is_index"] and not self.cleaned_data["llm_provider"]:
             raise forms.ValidationError("This field is required when creating an index.")
         return self.cleaned_data["llm_provider"]
-
-
-class FileForm(forms.ModelForm):
-    class Meta:
-        model = File
-        fields = ["name", "summary", "file"]
-        help_texts = {
-            "summary": "This is only needed when the file will not be used for RAG",
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["file"].widget.attrs.update({"class": "file-input"})
