+++ apps/documents/views.py
@@ -11,9 +11,9 @@ from django.views.decorators.http import require_POST
 from django.views.generic import CreateView, TemplateView, UpdateView
 from django_tables2 import SingleTableView
 
-from apps.documents.forms import CollectionForm, FileForm
+from apps.documents.forms import CollectionForm
 from apps.documents.models import Collection, CollectionFile, FileStatus
-from apps.documents.tables import CollectionsTable, FilesTable
+from apps.documents.tables import CollectionsTable
 from apps.files.models import File
 from apps.generics.chips import Chip
 from apps.teams.decorators import login_and_team_required
@@ -178,97 +178,4 @@ class DeleteCollection(LoginAndTeamRequiredMixin, View):
         else:
             collection.archive()
             messages.success(request, "Collection deleted")
-            return HttpResponse()
-
-
-class FileHome(LoginAndTeamRequiredMixin, TemplateView):
-    template_name = "generic/object_home.html"
-
-    def get_context_data(self, team_slug: str, **kwargs):
-        return {
-            "active_tab": "files",
-            "title": "Files",
-            "new_object_url": reverse("documents:file_new", args=[team_slug]),
-            "table_url": reverse("documents:file_table", args=[team_slug]),
-            "enable_search": True,
-        }
-
-
-class FileTableView(LoginAndTeamRequiredMixin, SingleTableView):
-    model = File
-    paginate_by = 25
-    table_class = FilesTable
-    template_name = "table/single_table.html"
-
-    def get_queryset(self):
-        queryset = File.objects.filter(team=self.request.team, is_version=False).order_by("-created_at")
-        if search := self.request.GET.get("search"):
-            queryset = similarity_search(queryset, search_phase=search, columns=["name", "summary"], score=0.1)
-        return queryset
-
-
-class CreateFile(LoginAndTeamRequiredMixin, CreateView):
-    template_name = "documents/file_form.html"
-    model = File
-    form_class = FileForm
-    permission_required = "files.add_file"
-    extra_context = {
-        "title": "Upload File",
-        "button_text": "Upload",
-        "active_tab": "files",
-        "form_attrs": {"enctype": "multipart/form-data"},
-    }
-
-    def get_success_url(self):
-        return reverse("documents:file_home", args=[self.request.team.slug])
-
-    def form_valid(self, form):
-        form.instance.team = self.request.team
-        response = super().form_valid(form)
-        messages.success(self.request, "File uploaded successfully")
-        return response
-
-
-class EditFile(LoginAndTeamRequiredMixin, UpdateView):
-    template_name = "documents/file_form.html"
-    model = File
-    form_class = FileForm
-    permission_required = "files.change_file"
-    extra_context = {
-        "title": "Edit File",
-        "button_text": "Update",
-        "active_tab": "files",
-        "form_attrs": {"enctype": "multipart/form-data"},
-    }
-
-    def get_queryset(self):
-        return File.objects.filter(team=self.request.team)
-
-    def get_success_url(self):
-        return reverse("documents:file_home", args=[self.request.team.slug])
-
-    def post(self, request, *args, **kwargs):
-        response = super().post(request, *args, **kwargs)
-        messages.success(request, "File updated successfully")
-        return response
-
-
-class DeleteFile(LoginAndTeamRequiredMixin, View):
-    permission_required = "files.delete_file"
-
-    def delete(self, request, team_slug: str, pk: int):
-        file = get_object_or_404(File, team__slug=team_slug, id=pk)
-
-        if collections := file.get_collection_references():
-            response = render_to_string(
-                "assistants/partials/referenced_objects.html",
-                context={
-                    "object_name": "file",
-                    "pipeline_nodes": [Chip(label=col.name, url=col.get_absolute_url()) for col in collections],
-                },
-            )
-            return HttpResponse(response, headers={"HX-Reswap": "none"}, status=400)
-        else:
-            file.archive()
-            messages.success(request, "File deleted")
             return HttpResponse()
