+++ templates/documents/file_form.html
@@ -4,7 +4,7 @@
 {% block breadcrumbs %}
   <div class="text-sm breadcrumbs" aria-label="breadcrumbs">
     <ul>
-      <li><a href="{% url 'documents:file_table' request.team.slug %}">Files</a></li>
+      <li><a href="{% url 'files:file_table' request.team.slug %}">Files</a></li>
       {% if object.id %}
         <li class="pg-breadcrumb-active" aria-current="page">Edit</li>
       {% else %}
